# openapi.yaml
openapi: 3.0.3
info:
  title: Template Builder API
  version: 1.0.0
  description: >
    Intent-driven, AI-native template builder and document generation API.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/registry/placeholders:
    get:
      summary: List placeholder registry entries
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: datasource_id
          schema:
            type: integer
        - in: query
          name: classification
          schema:
            type: string
      responses:
        "200":
          description: List of placeholders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaceholderRegistryEntry"
    post:
      summary: Create a placeholder registry entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceholderRegistryCreateRequest"
      responses:
        "201":
          description: Placeholder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceholderRegistryEntry"

  # --------------------------------------------------------------------
  # Runtime Document Generation
  # --------------------------------------------------------------------
...
components:
  schemas:
...
    PlaceholderRegistryEntry:
      type: object
      properties:
        registry_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        generation_mode:
          type: string
          enum: [manual_sql, llm_prompt]
        prompt:
          type: string
        sql_text:
          type: string
        datasource_id:
          type: integer
        value_type:
          type: string
        cardinality:
          type: string
        classification:
          type: string
        format_json:
          type: object
        sample_value:
          type: string
        metadata:
          type: object
        created_by:
          type: string
    PlaceholderRegistryCreateRequest:
      type: object
      required: [name, generation_mode, datasource_id]
      properties:
        name:
          type: string
        description:
          type: string
        generation_mode:
          type: string
          enum: [manual_sql, llm_prompt]
        prompt:
          type: string
        sql_text:
          type: string
        datasource_id:
          type: integer
        value_type:
          type: string
          default: string
        cardinality:
          type: string
          default: scalar
        classification:
          type: string
          default: internal
        format_json:
          type: object
        sample_value:
          type: string
        metadata:
          type: object
...